- flipping the condition to include for ajax
Before it was set to && which means it had to have both conditions
correct to make it work.

Now we've got || which means if we're NOT in the WordPress admin OR
if we're DOING_AJAX we setup the MVC.

Not sure why my original testing went well, maybe I was just being
hasty.